move timeout validation to utilities
