+++ uds/utilities/__init__.py
@@ -102,3 +102,4 @@ from .custom_warnings import (
     ValueWarning,
 )
 from .enums import ByteEnum, Endianness, ExtendableEnum, NibbleEnum, ValidatedEnum
+from .helpers import validate_timeout
